+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next
 
 * `fixed` Error verbosity when failing to read ACME directory
+* `fixed` Correctly recognize `ready` and `processing` states - [RFC 8555 Section 7.1.6](https://tools.ietf.org/html/rfc8555#section-7.1.6)
 
 
 ## v4.1.4 (2021-12-23)